Start util file for common functions.
